requirements.txt
Click==7.0
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
dlib==19.8.1
face-recognition==1.2.3
face-recognition-models==0.3.0
gunicorn==19.9.0
numpy==1.16.2
Pillow==5.4.1
psycopg2==2.7.7
pytz==2018.9
whitenoise==4.1.2
